+++ src/components/paymentPage/payTitle/PayTitle.tsx
@@ -1,0 +1,6 @@
+import * as t from './PayTitle.style'
+const PayTitle = () => {
+  return <t.H1tag>결제하기</t.H1tag>
+}
+
+export default PayTitle
