+++ src/components/paymentPage/payUserInput/PayUserInput.style.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const DivArea = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+`
+export const Div = styled.div`
+  width: 49%;
+`
