+++ src/components/paymentPage/pdtInfo/PdtInfo.tsx
@@ -1,0 +1,22 @@
+import * as t from './PdtInfo.style'
+
+const PdtInfo = () => {
+  return (
+    <>
+      <t.ListArea>
+        <t.PdArea>
+          <t.PdInfoDiv>
+            <t.ProductImg />
+            <t.ProductInfo>
+              <t.ProductName>[지구샵] 스트롱 고체치약(쿨민트향)</t.ProductName>
+              <t.ProducOption>틴케이스 - 1개</t.ProducOption>
+              <t.ProducPrice>1,000원 / 3 개</t.ProducPrice>
+            </t.ProductInfo>
+          </t.PdInfoDiv>
+        </t.PdArea>
+      </t.ListArea>
+    </>
+  )
+}
+
+export default PdtInfo
