+++ src/containers/paymentPage/OrderPay.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
 
-import PayRadioBtn from '../../components/paymentPage/PayRadioBtn'
-import PaySummary from '../../components/paymentPage/PaySummary'
+import PayRadioBtn from '../../components/paymentPage/payRadioBtn/PayRadioBtn'
+import PaySummary from '../../components/paymentPage/paySummary/PaySummary'
 import PayAgree from '../../components/paymentPage/payAgree/PayAgree'
 import { OrderInfoDiv } from './OrderPList'
 import { MainButton } from '../../elements/Buttons'
